Add TODO for pickup ingredient conversion
Pickup is not passing antecedents to steps, which then causes a failure during
ingredient conversion. This isn't a hard-stop, as this is only used for
codified metadata, but it does mean that we will need to rewrite a little to
generate codified metadata from the cache for the earliest pickup step in the
process.