Meta revamp (#8)
* Remove role-based sub-folders

These folders introduced unnecessary complication when the role of steps
shifted; caches that were previously a "product" may become an "intermediary",
but we still want to treat the cache as valid. This requires a consistent path
be used, regardless of role.

* Remove volatile elements from timing

Timing changes each time metadata is executed. This will need to be handled
in a separate folder, along with other volatile metadata to allow consistent
generation of metadata.

* Remove execute from step init

This allows us to inspect some attributes prior to execution.

* Alter cache check to use constructed attribute

* Add cache note

* Adding product lister

For use in earliest-available

* Abstract role to recipe layer

Role of a Step is determined with the full context of a recipe, so there is
no value in assigning roles directly to a Step or preserving them in metadata.

* Remove role attribute references

* Switch role lookup source

* Checkpoint

* Working through step stuff

* Fixed ingredient calls

* Made stepper able to fire before execution

* Stub for pickup method

* Fix default keep

* Checkpoint

* Disable broken function

* Refactoring metadata

* Working on lineage

* Working out new metadata class structure

* Working out appropriate lineage fingerprinting

* Correct metadata parent conversion

* This might be working now

* Shorten fingerprinting

* Prepping for step contstructor with codified meta

* Working out codified metadata

* Working on metadata test coverage

* I don't really know at this point

* Adding some notes on what needs to be tested

This is still a hot mess.

* Trying to correct this

It looks like I'm actually going to have to write good metadata
in these tests, which is really hard to do with the tooling...

* Working on JSON schema definition

* Saving example validation to main call of meta

* Remove unnecessary placeholder for id

Since we're using this schema in a python implementation anyways, and
overwriting the placeholder value, it's cleaner to just omit it.

* Make codified node required

That is the core of this whole thing, after all.

* Work on assertions with JSON Schema

Since the JSON Schema validation can handle a bulk of the work related to
field validation, the python will focus instead on the validation of
relationships between values. Most specifically, the fields used between nodes
of lineage declarations.

* Trying to work new metadata construction

* Working on Step metadata passing

* Fixing schema

* Fixing metadata output

* Holy crap it worked

* Some error messages help

* Docs and stuff

* Incorporate pickup step method

This identifies which steps need to be executed in order to "pick up" the
recipe from the latest possible point in the execution, based upon the contents
of the cache.

* Debugging

* Add TODO for pickup ingredient conversion

Pickup is not passing antecedents to steps, which then causes a failure during
ingredient conversion. This isn't a hard-stop, as this is only used for
codified metadata, but it does mean that we will need to rewrite a little to
generate codified metadata from the cache for the earliest pickup step in the
process.

* Working out print statements

* Working through caching issues

* Working through caching issues metadata schema stuff

* Working on new python handling of schema logic

* Finishing up python metadata schema methods

* Finally have a reasonable test strategy for metadata

* Flushing out cases

* Adding some notes

* Refactoring metadata expectations

* Doing it

* Strip schema declaration

This was apparently a problem (found a bad example on the internet,
apparently), and it was causing the jsonschema python package to throw
deprecation warnings.

* Working up some test stuff

* Add valid minimal example

* Remove unnecessary validation in metadata object handler

* Catching incorrect fingerprints now

* Dealing with fingerprint issues in test cases

* Fix sorting issues in full test case

* Remove broken, unused tests

* Expand step output typehint

So that Path arguments work in IDE even when output is implicitly defined

* These tests need a lot of work

* Fix return error test

It was calling instructions directly, rather than through the execute method,
which performs the error handling.

* Fix output requirement for keep

* Fix case for cache testing

* Fix sub-folder output test

* Fix order execution test

* Metadata docs

* Working on docstrings

* Cleaning up metadata docs

* I have no words

* Documenting pickup test

* Flushing out metadata tests and documentation

* Remove plotly stuff

It was saved off into it's own branch if I want to dig it up later

* Clean up warnings on schema

* Remove unused exception customizations

* Working on Step coverage

* Remove unnecessary (and buggy) method

* Completing step test coverage

Among other things...